pyproject.toml
[tool.poetry]
name = "adjango-template"
version = "2.0.0"
description = ""
authors = ["Artasov <ivanhvalevskey@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"

# Tier 1
Django = "5.2.5"
adjango = "0.6.0"
djangorestframework = "3.16.1"
adrf = "0.1.9"
celery = "5.5.3"
daphne = "4.2.1"
channels = "4.3.1"

# Tier 2
django-redis = "6.0.0"
django-endpoints = "0.0.3"
django-extensions = "4.1"
django-polymorphic = "4.1.0"
django-celery-beat = "2.8.1"
django-cachalot = "2.8.0"
django-silk = "5.4.1"
django-imagekit = "5.0.0"
django-user-agents = "0.4.0"
django-object-actions = "5.0.0"
django-import-export = { version = "4.3.8", extras = ["all"] }

pytest-asyncio = "0.26.0"
pytest-django = "4.11.1"
drf-yasg = "1.21.10"
redis = "6.4.0"
channels-redis = "4.3.0"
twisted = { version = "25.5.0", extras = ["http2", "tls"] }
polib = "1.2.0"

aiohttp = "3.11.18"
chardet = "5.2.0"
python-dotenv = "1.1.1"

Pillow = "11.0.0"
psycopg2-binary = "2.9.10"
requests = "2.32.4"
setuptools = "80.4.0"
psutil = "7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
